// Frontend/sakai-ng-master/src/app/biblioteca/interfaces/reportes/intranet-visita.ts
export interface IntranetVisitaDTO {
    tipoUsuario: string;
    subTipoUsuario: string;
    programa: string;
    pacpro: string;
    idVisitaBibVir: string;
    idSituacionAlumno: string;
    idBibVir: string;
    horaSalida: string;
    horaIngreso: string;
    flgUsuario: string;
    fechaRegistro: string;
    codigoUsuario: string;
    codigoSede: string;
    codigoEspecialidad: string;
    ciclo: string;
    totalVisitas: number;
}